0.2.28 does more sanity checks on recipe files, more sanity checks on memory accesses, and generally replaces a lot of ints with longs in order to handle large memory allocations better

+++ cuda_util.h
@@ -23,26 +23,26 @@ public:
 
 // Helper to calculate a 2d row-major index, ie for:
 //   arr[a][b]
-__host__ __device__ inline int index2d(int a, int b, int b_end) {
+__host__ __device__ inline long index2d(long a, long b, long b_end) {
   return a * b_end + b;
 }
 
 // Helper to calculate a 3d row-major index, ie for:
 //   arr[a][b][c]
-__host__ __device__ inline int index3d(int a, int b, int b_end, int c, int c_end) {
+__host__ __device__ inline long index3d(long a, long b, long b_end, long c, long c_end) {
   return index2d(a, b, b_end) * c_end + c;
 }
 
 // Helper to calculate a 4d row-major index, ie for:
 //   arr[a][b][c][d]
-__host__ __device__ inline int index4d(int a, int b, int b_end, int c, int c_end, int d, int d_end) {
+__host__ __device__ inline long index4d(long a, long b, long b_end, long c, long c_end, long d, long d_end) {
   return index3d(a, b, b_end, c, c_end) * d_end + d;
 }
 
 // Helper to calculate a 5d row-major index, ie for:
 //   arr[a][b][c][d][e]
-__host__ __device__ inline int index5d(int a, int b, int b_end, int c, int c_end, int d, int d_end,
-                                int e, int e_end) {
+__host__ __device__ inline long index5d(long a, long b, long b_end, long c, long c_end, long d, long d_end,
+                                long e, long e_end) {
   return index4d(a, b, b_end, c, c_end, d, d_end) * e_end + e;
 }
